.claude/skills/world-builder/_lib/requirements.txt
networkx>=3.0
chromadb>=0.4.0
PyYAML>=6.0
google-genai>=1.0.0
openai>=1.0.0
